Migrate to pydantic, add observation data superclass
