+++ sdks/python/src/opik/api_objects/span/span_client.py
@@ -1,5 +1,6 @@
 import datetime
 import logging
+
 from typing import Any, Dict, List, Optional, Union
 
 from opik import datetime_helpers, id_helpers, llm_usage
